Refactor decoder output format
The natural goal in the end is to have two converters:

	takum -> logarithmic value
	linear takum -> floating point value

While we match the first goal, the second was not fully obtained as
we simply extracted the fraction bits in the linear case. However, the
fraction bits as interpreted within the posits context are for a
floating-point representation of the form

	[(1 - 3 * S) + f] * 2^e

If S=1 we thus have the form

	[-2 + f] * 2^e

While this makes a lot of sense to preserve monotonicity, conversion to
IEEE floating-point requires the representation

	(-1)^S * [1 + g] * 2^e,

with g in [0,1), which is obtained via f = 1-g. However, this is not as
straightforward as it may seem, given f in [0,1) can be zero, thus
yielding g = 1, which is not in [0,1). The trick is to borrow a factor 2
from the exponent, but this quickly gets complicated.

There is a simpler solution, which also streamlines the code: The math we
do in the logarithmic case actually is everything we need
for the linear case, given the logarithmic value properly reflects the
use of the borrowing in the 'exponent' when subtracting 1 from an
all-zero F. Thus we can streamline the code to have the decoder.vhd for
the default logarithmic case and then the decoder_linear.vhd simply
calls into the decoder entity and maps the logarithmic value bits
accordingly.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>