Revert "Refactor encoder output format"
After more deliberation, while it makes sense to appease the conversion
to floating-point, it makes more sense to go for a different internal
representation as discussed by Murillo et al. in 'Comparing Different
Decodings for Posit Arithmetic' which does not require the calculation
of the two's-complement. This will be further developed in subsequent
commits.

This reverts commit d550d08831cb574841c8b2f238f74d639336e80f.